Move OutputFiles to GenericJob
