+++ pyiron_base/jobs/flex/executablecontainer.py
@@ -1,4 +1,3 @@
-import os
 import cloudpickle
 import numpy as np
 from pyiron_base.jobs.job.template import TemplateJob
@@ -44,7 +43,6 @@ class ExecutableContainerJob(TemplateJob):
         super().__init__(project, job_name)
         self._write_input_funct = None
         self._collect_output_funct = None
-        self.output_files = OutputFiles(working_directory=self.working_directory)
 
     def set_job_type(
         self,
@@ -109,21 +107,3 @@ class ExecutableContainerJob(TemplateJob):
             self._collect_output_funct = cloudpickle.loads(
                 self.project_hdf5["collect_output_function"]
             )
-
-
-class OutputFiles:
-    def __init__(self, working_directory):
-        self._working_directory = working_directory
-
-    def __dir__(self):
-        return list(self._get_file_convert_dict().keys())
-
-    def _get_file_convert_dict(self):
-        return {f.replace(".", "_"): f for f in os.listdir(self._working_directory)}
-
-    def __getattr__(self, attr):
-        convert_dict = self._get_file_convert_dict()
-        if attr in convert_dict.keys():
-            return os.path.join(self._working_directory, convert_dict[attr])
-        else:
-            raise AttributeError()
