+++ tests/flex/test_executablecontainer.py
@@ -81,6 +81,24 @@ class TestExecutableContainer(TestWithProject):
         self.assertEqual(job_reload.input["energy"], energy_value)
         self.assertEqual(job_reload.output["energy"], energy_value)
 
+    def test_job_output_files(self):
+        create_catjob = create_job_factory(
+            write_input_funct=write_input,
+            collect_output_funct=collect_output,
+            default_input_dict={"energy": 1.0},
+            executable_str="cat input_file > output_file",
+        )
+        job = create_catjob(
+            project=ProjectHDFio(project=self.project, file_name="any.h5", h5_path=None, mode=None),
+            job_name="job_output_files"
+        )
+        job.run()
+        self.assertEqual(dir(job.output_files), ['error_out', 'input_file', 'output_file'])
+        self.assertEqual(
+            job.output_files.error_out,
+            os.path.abspath(os.path.join(__file__, "..", "test_executablecontainer", "job_output_files_hdf5", "job_output_files", "error.out"))
+        )
+
     def test_create_job_factory_typeerror(self):
         create_catjob = create_job_factory(
             write_input_funct=write_input,
